+++ sayn/database/__init__.py
@@ -22,6 +22,7 @@ class DDL(BaseModel):
         primary: Optional[bool] = False
         not_null: Optional[bool] = False
         unique: Optional[bool] = False
+        dst_name: Optional[str]
 
     class Index(BaseModel):
         columns: conlist(str, min_items=1)
@@ -271,9 +272,10 @@ class Database:
         """
         with self.engine.connect().execution_options(stream_results=True) as connection:
             res = connection.execute(query, **params)
+            fields = res.keys()
 
             for record in res:
-                yield dict(zip(res.keys(), record))
+                yield dict(zip(fields, record))
 
     def _load_data_batch(self, table, data, schema):
         """Implements the load of a single data batch for `load_data`.
@@ -315,6 +317,9 @@ class Database:
               (True) or new records are to be appended to the existing table (default)
             ddl (dict): An optional ddl specification in the same format as used
               in autosql and copy tasks
+
+        Returns:
+            int: Number of records loaded
         """
         batch_size = batch_size or self.max_batch_rows
         buffer = list()
@@ -333,6 +338,7 @@ class Database:
         check_create = True
         table_exists_prior_load = self._table_exists(table, schema)
 
+        records_loaded = 0
         for i, record in enumerate(data):
             if check_create and (replace or not table_exists_prior_load):
                 # Create the table if required
@@ -351,12 +357,16 @@ class Database:
 
             if i % batch_size == 0 and len(buffer) > 0:
                 self._load_data_batch(table, buffer, schema)
+                records_loaded += len(buffer)
                 buffer = list()
 
             buffer.append(record)
 
         if len(buffer) > 0:
             self._load_data_batch(table, buffer, schema)
+            records_loaded += len(buffer)
+
+        return records_loaded
 
     def _get_table(self, table, schema):
         """Create a SQLAlchemy Table object.