+++ sayn/tasks/copy.py
@@ -60,6 +60,8 @@ class Config(BaseModel):
     ddl: Optional[Dict[str, Any]]
     delete_key: Optional[str]
     incremental_key: Optional[str]
+    max_merge_rows: Optional[int]
+    max_batch_rows: Optional[int]
 
     @validator("incremental_key", always=True)
     def incremental_validation(cls, v, values):
@@ -67,6 +69,13 @@ class Config(BaseModel):
             raise ValueError(
                 'Incremental copy requires both "delete_key" and "incremental_key"'
             )
+
+        return v
+
+    @validator("max_merge_rows")
+    def merge_batch_size_val(cls, v, values):
+        if values.get("delete_key") is None:
+            raise ValueError("max_merge_rows is only applicable to incremental copy")
 
         return v
 
@@ -160,25 +169,58 @@ class CopyTask(SqlTask):
         )
 
         self.delete_key = self.config.delete_key
-        self.incremental_key = self.config.incremental_key
-
+        self.src_incremental_key = self.config.incremental_key
+        self.dst_incremental_key = self.config.incremental_key
+        self.max_merge_rows = self.config.max_merge_rows
+        self.max_batch_rows = self.config.max_batch_rows
         self.is_full_load = self.run_arguments["full_load"] or self.delete_key is None
 
         result = self.target_db._validate_ddl(self.config.ddl)
         if result.is_ok:
             self.ddl = result.value
+
+            # Check if the incremental_key in the destination needs renaming
+            if self.dst_incremental_key is not None and len(self.ddl["columns"]) > 0:
+                columns_dict = {
+                    c["name"]: c["dst_name"] or c["name"] for c in self.ddl["columns"]
+                }
+                self.dst_incremental_key = columns_dict[self.src_incremental_key]
         else:
             return result
 
         return Ok()
 
     def compile(self):
-        return self.execute(False, self.run_arguments["debug"])
+        result = self.get_columns()
+        if result.is_err:
+            return result
+
+        return self.execute(False, self.run_arguments["debug"], self.is_full_load)
 
     def run(self):
-        return self.execute(True, self.run_arguments["debug"])
+        result = self.get_columns()
+        if result.is_err:
+            return result
 
-    def execute(self, execute, debug):
+        if self.max_merge_rows is not None:
+            result = self.execute(
+                True,
+                self.run_arguments["debug"],
+                self.is_full_load,
+                self.max_merge_rows,
+            )
+            if result.is_err:
+                return result
+            for _ in range(100):
+                if result.is_err or result.value < self.max_merge_rows:
+                    break
+                result = self.execute(True, False, False, self.max_merge_rows)
+        else:
+            result = self.execute(True, self.run_arguments["debug"], self.is_full_load)
+
+        return result
+
+    def execute(self, execute, debug, is_full_load, limit=None):
         # Introspect target
         self.target_table_exists = self.target_db._table_exists(self.table, self.schema)
 
@@ -186,7 +228,7 @@ class CopyTask(SqlTask):
         if self.target_table_exists:
             load_table = self.tmp_table
             load_schema = self.tmp_schema
-            if self.is_full_load or self.incremental_key is None:
+            if is_full_load or self.src_incremental_key is None:
                 steps.append("Move Table")
             else:
                 steps.append("Merge Tables")
@@ -197,11 +239,7 @@ class CopyTask(SqlTask):
         self.set_run_steps(steps)
 
         with self.step("Prepare Load"):
-            result = self.get_columns()
-            if result.is_err:
-                return result
-
-            result = self.get_read_query(execute, debug)
+            result = self.get_read_query(execute, debug, is_full_load, limit)
             if result.is_err:
                 return result
             else:
@@ -219,9 +257,15 @@ class CopyTask(SqlTask):
                     return Exc(e)
 
         with self.step("Load Data"):
+            n_records = 0
             if execute:
                 data_iter = self.source_db._read_data_stream(get_data_query)
-                self.target_db.load_data(load_table, data_iter, schema=load_schema)
+                n_records = self.target_db.load_data(
+                    load_table,
+                    data_iter,
+                    schema=load_schema,
+                    batch_size=self.max_batch_rows,
+                )
 
         # Final step
         final_step = steps[-1]
@@ -257,15 +301,13 @@ class CopyTask(SqlTask):
                     except Exception as e:
                         return Exc(e)
 
-        return Ok()
+        return Ok(n_records)
 
     def get_columns(self):
         # We get the source table definition
         source_table_def = self.source_db._get_table(
             self.source_table,
             self.source_schema,
-            # columns=[c["name"] for c in self.ddl["columns"]],
-            # required_existing=True,
         )
         if source_table_def is None:
             return Err(
@@ -308,36 +350,48 @@ class CopyTask(SqlTask):
 
             else:
                 # In any other case, we use the source
-                self.ddl["columns"] = [
-                    {"name": c.name, "type": self.target_db._py2sqa(c.type.python_type)}
-                    for c in self.source_table_def.columns
-                ]
+                for c in self.source_table_def.columns:
+                    try:
+                        col_type = self.target_db._py2sqa(c.type.python_type)
+                    except:
+                        col_type = c.type.compile(self.target_db.engine.dialect)
+                    self.ddl["columns"].append({"name": c.name, "type": col_type})
         else:
             # Fill up column types from the source table
             for col in self.ddl["columns"]:
                 if col.get("name") not in self.source_table_def.columns:
                     return Err(
                         "database_error",
-                        "source_table_missing_column",
+                        "source_table_missing_columns",
                         db=self.source_db.name,
                         table=self.source_table,
                         schema=self.source_schema,
                         column=col.get("name"),
                     )
 
-                if "type" not in col:
-                    col["type"] = self.target_db._py2sqa(
-                        self.source_table_def.columns[col["name"]].type.python_type
-                    )
+                if col.get("type") is None:
+                    try:
+                        col["type"] = self.source_table_def.columns[
+                            col["name"]
+                        ].type.compile(self.target_db.engine.dialect)
+                    except:
+                        col["type"] = self.target_db._py2sqa(
+                            self.source_table_def.columns[col["name"]].type.python_type
+                        )
+
+        for col in self.ddl["columns"]:
+            col["src_name"] = col["name"]
+            if col.get("dst_name") is not None:
+                col["name"] = col["dst_name"]
 
         return Ok()
 
-    def get_read_query(self, execute, debug):
+    def get_read_query(self, execute, debug, is_full_load, limit=None):
         # Get the incremental value
         last_incremental_value_query = (
-            f"SELECT MAX({self.incremental_key}) AS value\n"
+            f"SELECT MAX({self.dst_incremental_key}) AS value\n"
             f"FROM {'' if self.schema is None else self.schema +'.'}{self.table}\n"
-            f"WHERE {self.incremental_key} IS NOT NULL"
+            f"WHERE {self.dst_incremental_key} IS NOT NULL"
         )
         if debug:
             self.write_compilation_output(
@@ -345,12 +399,17 @@ class CopyTask(SqlTask):
             )
 
         get_data_query = select(
-            columns=[column(c["name"]) for c in self.ddl["columns"]],
+            columns=[
+                column(c["src_name"]).label(c["name"])
+                if c["src_name"] != c["name"]
+                else column(c["src_name"])
+                for c in self.ddl["columns"]
+            ],
             from_obj=self.source_table_def,
         )
         last_incremental_value = None
 
-        if not self.is_full_load and self.target_table_exists:
+        if not is_full_load and self.target_table_exists:
             if execute:
                 res = self.target_db.read_data(last_incremental_value_query)
                 if len(res) == 1:
@@ -362,12 +421,24 @@ class CopyTask(SqlTask):
         if last_incremental_value is not None:
             get_data_query = get_data_query.where(
                 or_(
-                    self.source_table_def.c[self.incremental_key].is_(None),
-                    self.source_table_def.c[self.incremental_key]
-                    > last_incremental_value,
+                    self.source_table_def.c[self.src_incremental_key].is_(None),
+                    self.source_table_def.c[self.src_incremental_key]
+                    >= last_incremental_value,
                 )
             )
+
+        if self.src_incremental_key is not None:
+            get_data_query = get_data_query.order_by(self.src_incremental_key)
+
+        if limit is not None:
+            get_data_query = get_data_query.limit(limit)
+
         if debug:
-            self.write_compilation_output(get_data_query, "get_data")
+            try:
+                q = get_data_query.compile(compile_kwargs={"literal_binds": True})
+            except:
+                # compilation can fail when using values like dates
+                q = str(get_data_query)
+            self.write_compilation_output(q, "get_data")
 
         return Ok(get_data_query)